update feature schema
